Remove reference to libcurl typecheck header
`SetDefaultValue()` takes exactly three arguments, but the type of the
third variable is not known at declaration time, so it is defined as a
varargs function, and a wrapper macro is used to enforce that only
exactly three arguments can be passed.

This is fine, but the code to implement this was copied from `curl.h`
and references an internal header (`typecheck-gcc.h`) which is
(presumably) not supposed to be accessed from external code.

The include statement is guarded by a `!defined(__cplusplus)` check,
so it is completely ignored in C++ code.

But when including `notifications.h` from C code, `typecheck-gcc.h`
will be included, and will cause an error because it can't be found.
Because `notifications.h` is not alongside the other CURL headers,
the correct include statement would be `<curl/typecheck-gcc.h>`.

But even that is problematic. From a quick skim of that file, it seems
to be an internal header which only defines CURL-specific macros that
operate on CURL-specific data structures. As far as I can tell, it is
completely useless for libnotifications.

So I've removed it and its surrounding guards. The actual
`SetDefaultValue()` macro which enforces the argument count is still
present, it is now just hidden behind one less guard.